Adjust tests; move some out of lib.rs
